requirements.txt
# Core RL & Environment
gymnasium>=0.29.0
mujoco>=3.0.0
stable-baselines3>=2.1.0
sb3-contrib>=2.1.0

# Object Detection
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data & Visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Documentation
jupyter>=1.0.0
